+++ wtframework/wtf/_devtools_/filetemplates/_test_template_.py
@@ -1,4 +1,7 @@
-content = \
+from six import u
+
+
+contents = u( \
 '''
 """
 
@@ -28,4 +31,4 @@ class {testname}(WTFBaseTest):
 if __name__ == "__main__":
     #import sys;sys.argv = ['', 'Test.testName']
     unittest.main()
-'''
+''')