+++ wtframework/wtf/config.py
@@ -15,7 +15,6 @@
 #    along with WTFramework.  If not, see <http://www.gnu.org/licenses/>.
 ##########################################################################
 
-
 from types import NoneType
 from wtframework.wtf.utils.project_utils import ProjectUtils
 import os
@@ -23,6 +22,8 @@ import re
 import yaml
 import time
 
+from six import u
+from wtframework.wtf import _wtflog
 
 class ConfigReader:
     '''
@@ -53,7 +54,7 @@ class ConfigReader:
                 for config in reversed(configs):
                     self.__load_config_file(config)
             elif not ConfigReader.ENV_VARS in os.environ:
-                print "Config file not specified.  Using config/defaults.yaml"
+                _wtflog.warning(u("Config file not specified.  Using config/defaults.yaml"))
                 self.__load_config_file(ConfigReader.DEFAULT_CONFIG_FILE)
             else:
                 # Read and load in all configs specified in reverse order
@@ -64,9 +65,9 @@ class ConfigReader:
                 
         except Exception as e:
             # Fall back to default.yaml file when no config settings are specified.
-            print "An error occurred while loading config file:", e
+            _wtflog.error(u("An error occurred while loading config file: %s"), e)
             raise e
-            
+
 
 
     class __NoDefaultSpecified__(object):
@@ -114,7 +115,7 @@ class ConfigReader:
                 pass
             
         if default_value == self.__NoDefaultSpecified__:
-            raise KeyError("Key '{0}' does not exist".format(key))
+            raise KeyError(u("Key '{0}' does not exist").format(key))
         else:
             return default_value
 
@@ -124,14 +125,14 @@ class ConfigReader:
             config_file_location = os.path.join(ProjectUtils.get_project_root(),
                                                 ConfigReader.CONFIG_LOCATION,
                                                 file_name + ConfigReader.CONFIG_EXT)
-            print u"locating config file:", config_file_location
+            _wtflog.debug(u("locating config file: %s"), config_file_location)
             config_yaml = open(config_file_location, 'r')
             dataMap = yaml.load(config_yaml)
             self._dataMaps.insert(0, dataMap)
             config_yaml.close()
         except Exception as e:
-            print u"Error loading config file " + file_name
-            raise ConfigFileReadError(u"Error reading config file " + file_name, e)
+            _wtflog.error(u("Error loading config file: %s"), file_name)
+            raise ConfigFileReadError(u("Error reading config file ") + file_name, e)
 
 
 class ConfigFileReadError(RuntimeError):